improve ard0 Serial and substr & bugfix
This does not fully resolve the issue where ard0 will sometimes become
unresponsive. Currently, it has been shown that this is unlikely a
purely Serial problem, or one with explicit use of the heap. It is
possible that this is a power supply issue or a library problem.

The patch has also only been applied to ard0 and not the other arduinos.